+++ paper.bib
@@ -21,6 +21,7 @@
   year      = {1983},
   edition   = {2},
   isbn      = {0-89859-268-2}
+}
 
  @Manual{caret,
     title = {caret: Classification and Regression Training},
@@ -37,3 +38,4 @@
     note = {R package version 0.1.2},
     url = {https://CRAN.R-project.org/package=recipes},
   }
+  
