### 2018-02-11 #579 @danielskatz
hmmm. I don't have any experience with R, and I'm much more of a computer scientist and computational scientist than a data scientist.  I really don't think I'm the right person to edit this one.
