### 2018-02-11 #579 @JohnMount
Perhaps one or more of @milesmcbain @earino @Fusionet24 @nhejazi @trallard could review?  I was just looking through for people with a combination of interests in R, statistics, data science, and data preparation.
