### 2018-02-11 #579 @JohnMount
@danielskatz sorry for the trouble.  Should I suggest another editor or is there some other way to go forward?  I just found some R people from the review list.
