### 2018-02-11 #579 @danielskatz
If you can suggest another editor, I certainly would appreciate it :)
