### 2018-02-11 #579 @JohnMount
Perhaps Thomas J. Leeper (@leeper).  A social scientist (they know their statistics) and CRAN contributor.
