### 2018-02-11 #579 @JohnMount
@arfon thank for helping with this.  I assume that I do not assign editors, but have put in some suggestions.  Please let me know if I need to do something.
