### 2018-02-11 #579 @earino
But instinctively this sounds like something I would enjoy.

On Sun, Feb 11, 2018, 3:27 PM Eduardo Arino de la Rubia <earino@gmail.com>
wrote:

> LOL hang on, let's make sure I have any idea what I'm doing. I mostly just
> asked if I was needed :-) what are the requirements for excellent reviewing?
>
> On Sun, Feb 11, 2018, 7:43 AM Nima Hejazi <notifications@github.com>
> wrote:
>
>> Looks like @earino <https://github.com/earino> has already stepped up,
>> but also happy to review should it become necessary.
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/579#issuecomment-364727244>,
>> or mute the thread
>> <https://github.com/notifications/unsubscribe-auth/AAAMukg6gWq57TR7Y_NlOt4ABCZKOP5eks5tTov_gaJpZM4SBKb->
>> .
>>
>
