### 2018-02-14 #584 @whedon
[REVIEW]: The vtreat R package: a statistically sound data processor for predictive modeling
**Submitting author:** @JohnMount (<a href="http://orcid.org/0000-0002-3696-2012">John Mount</a>)
**Repository:** <a href="https://github.com/WinVector/vtreat/" target ="_blank">https://github.com/WinVector/vtreat/</a>
**Version:** 1.0.2
