### 2018-02-14 #584 @JohnMount
Just a note: we have included testthat tests in the package (an R way of automating tests).
