### 2018-02-16 bba76a8
get some timings
