### 2018-02-16 b4fcff9
work on parallelizing indicator formation
