### 2018-02-16 59c6c24
minor clean up
