+++ README.md
@@ -114,12 +114,12 @@ Trivial example:
 ``` r
 library("vtreat")
 packageVersion("vtreat")
- #  [1] '1.0.2'
+ #  [1] '1.0.3'
 citation('vtreat')
  #  
  #  To cite package 'vtreat' in publications use:
  #  
- #    John Mount and Nina Zumel (2017). vtreat: A Statistically Sound
+ #    John Mount and Nina Zumel (2018). vtreat: A Statistically Sound
  #    'data.frame' Processor/Conditioner.
  #    https://github.com/WinVector/vtreat/,
  #    https://winvector.github.io/vtreat/.
@@ -129,7 +129,7 @@ citation('vtreat')
  #    @Manual{,
  #      title = {vtreat: A Statistically Sound 'data.frame' Processor/Conditioner},
  #      author = {John Mount and Nina Zumel},
- #      year = {2017},
+ #      year = {2018},
  #      note = {https://github.com/WinVector/vtreat/, https://winvector.github.io/vtreat/},
  #    }
 
@@ -148,8 +148,8 @@ print(treatmentsC$scoreFrame[,c('origName', 'varName', 'code', 'rsq', 'sig', 'ex
  #  1        x  x_lev_NA   lev 0.296065432 0.09248399                 0
  #  2        x x_lev_x.a   lev 0.130005705 0.26490379                 0
  #  3        x x_lev_x.b   lev 0.006067337 0.80967242                 0
- #  4        x    x_catP  catP 0.111456141 0.30194137                 2
- #  5        x    x_catB  catB 0.115273608 0.29380616                 2
+ #  4        x    x_catP  catP 0.130498074 0.26400089                 2
+ #  5        x    x_catB  catB 0.030345745 0.59013918                 2
  #  6        z   z_clean clean 0.237601767 0.13176020                 0
  #  7        z   z_isBAD isBAD 0.296065432 0.09248399                 0
 
@@ -195,9 +195,9 @@ print(treatmentsN$scoreFrame[,c('origName', 'varName', 'code', 'rsq', 'sig', 'ex
  #  1        x  x_lev_NA   lev 3.333333e-01 0.1339746                 0
  #  2        x x_lev_x.a   lev 2.500000e-01 0.2070312                 0
  #  3        x x_lev_x.b   lev 1.110223e-16 1.0000000                 0
- #  4        x    x_catP  catP 3.558824e-01 0.1184999                 2
- #  5        x    x_catN  catN 2.131202e-02 0.7301398                 2
- #  6        x    x_catD  catD 4.512437e-02 0.6135229                 2
+ #  4        x    x_catP  catP 1.304740e-01 0.3793327                 2
+ #  5        x    x_catN  catN 2.616674e-01 0.1950536                 2
+ #  6        x    x_catD  catD 9.705853e-02 0.4525551                 2
  #  7        z   z_clean clean 2.880952e-01 0.1701892                 0
  #  8        z   z_isBAD isBAD 3.333333e-01 0.1339746                 0
 dTrainNTreated <- prepare(treatmentsN,dTrainN,pruneSig=1.0,scale=TRUE)
