### 2018-02-16 efae008
try to get more parellism on indicator variables
