### 2018-02-16 e8e3b63
better per-level parallism on indicator variable creation. need to work on the scoring
