+++ extras/PerfTests.md
@@ -6,6 +6,8 @@ Win-Vector LLC
 ``` r
 library("vtreat")
 
+useParallel <- FALSE
+
 mkEx <- function(n_rows, 
                  n_cat_columns, n_num_columns, n_irrel_columns,
                  n_cat_levels_a, n_cat_levels_b) {
@@ -49,9 +51,12 @@ mkEx <- function(n_rows,
   d
 }
 
-ncores <- parallel::detectCores()
-parallelCluster <- parallel::makeCluster(ncores)
-n_rows <- 500000
+parallelCluster <- NULL
+if(useParallel) {
+  ncores <- parallel::detectCores()
+  parallelCluster <- parallel::makeCluster(ncores)
+}
+n_rows <- 100000
 ```
 
 Get a base timing of a moderately large task.
@@ -77,30 +82,30 @@ system.time(
 ```
 
     ##    user  system elapsed 
-    ##  12.646   1.693  66.073
+    ##  18.793   3.004  21.878
 
 ``` r
 knitr::kable(tplan$treatments$scoreFrame)
 ```
 
-| varName                       | varMoves |      rsq|        sig| needsSplit |  extraModelDegrees| origName    | code  |
-|:------------------------------|:---------|--------:|----------:|:-----------|------------------:|:------------|:------|
-| var\_cat\_1\_lev\_x.lev\_a\_1 | TRUE     |  3.5e-06|  0.1184941| FALSE      |                  0| var\_cat\_1 | lev   |
-| var\_cat\_1\_lev\_x.lev\_a\_2 | TRUE     |  1.3e-06|  0.3489834| FALSE      |                  0| var\_cat\_1 | lev   |
-| var\_cat\_1\_lev\_x.lev\_a\_3 | TRUE     |  4.0e-07|  0.6002284| FALSE      |                  0| var\_cat\_1 | lev   |
-| var\_cat\_1\_lev\_x.lev\_a\_4 | TRUE     |  1.0e-07|  0.7721816| FALSE      |                  0| var\_cat\_1 | lev   |
-| var\_cat\_1\_lev\_x.lev\_a\_5 | TRUE     |  7.4e-06|  0.0235845| FALSE      |                  0| var\_cat\_1 | lev   |
-| var\_cat\_1\_catP             | TRUE     |  8.0e-07|  0.4496472| TRUE       |                  4| var\_cat\_1 | catP  |
-| var\_cat\_1\_catB             | TRUE     |  1.4e-06|  0.3181307| TRUE       |                  4| var\_cat\_1 | catB  |
-| var\_cat\_2\_lev\_x.lev\_a\_1 | TRUE     |  5.0e-07|  0.5401118| FALSE      |                  0| var\_cat\_2 | lev   |
-| var\_cat\_2\_lev\_x.lev\_a\_2 | TRUE     |  5.0e-07|  0.5543446| FALSE      |                  0| var\_cat\_2 | lev   |
-| var\_cat\_2\_lev\_x.lev\_a\_3 | TRUE     |  1.5e-06|  0.3035759| FALSE      |                  0| var\_cat\_2 | lev   |
-| var\_cat\_2\_lev\_x.lev\_a\_4 | TRUE     |  2.0e-06|  0.2371430| FALSE      |                  0| var\_cat\_2 | lev   |
-| var\_cat\_2\_lev\_x.lev\_a\_5 | TRUE     |  2.7e-06|  0.1744633| FALSE      |                  0| var\_cat\_2 | lev   |
-| var\_cat\_2\_catP             | TRUE     |  1.0e-07|  0.8296674| TRUE       |                  4| var\_cat\_2 | catP  |
[…]
