+++ extras/PerfTests.Rmd
@@ -8,6 +8,8 @@ output: github_document
 ```{r}
 library("vtreat")
 
+useParallel <- FALSE
+
 mkEx <- function(n_rows, 
                  n_cat_columns, n_num_columns, n_irrel_columns,
                  n_cat_levels_a, n_cat_levels_b) {
@@ -51,9 +53,12 @@ mkEx <- function(n_rows,
   d
 }
 
-ncores <- parallel::detectCores()
-parallelCluster <- parallel::makeCluster(ncores)
-n_rows <- 500000
+parallelCluster <- NULL
+if(useParallel) {
+  ncores <- parallel::detectCores()
+  parallelCluster <- parallel::makeCluster(ncores)
+}
+n_rows <- 100000
 ```
 
 Get a base timing of a moderately large task.
@@ -157,7 +162,10 @@ knitr::kable(tplan$treatments$scoreFrame)
 
 
 ```{r cleanup}
-parallel::stopCluster(parallelCluster)
+if(!is.null(parallelCluster)) {
+  parallel::stopCluster(parallelCluster)
+  parallelCluster <- NULL
+}
 ```
 
 

[…]
