### 2018-02-16 52c6e5a
should be have a better grain of parallelism now
