+++ README.md
@@ -144,14 +144,14 @@ dTestC <- data.frame(x=c('a','b','c',NA),z=c(10,20,30,NA))
 treatmentsC <- designTreatmentsC(dTrainC,colnames(dTrainC),'y',TRUE,
                                  verbose=FALSE)
 print(treatmentsC$scoreFrame[,c('origName', 'varName', 'code', 'rsq', 'sig', 'extraModelDegrees')])
- #    origName   varName  code          rsq        sig extraModelDegrees
- #  1        x  x_lev_NA   lev 2.960654e-01 0.09248399                 0
- #  2        x x_lev_x.a   lev 1.300057e-01 0.26490379                 0
- #  3        x x_lev_x.b   lev 6.067337e-03 0.80967242                 0
- #  4        x    x_catP  catP 1.559780e-01 0.22202097                 2
- #  5        x    x_catB  catB 1.177202e-05 0.99153548                 2
- #  6        z   z_clean clean 2.376018e-01 0.13176020                 0
- #  7        z   z_isBAD isBAD 2.960654e-01 0.09248399                 0
+ #    origName   varName  code         rsq        sig extraModelDegrees
+ #  1        x    x_catP  catP 0.111456141 0.30194137                 2
+ #  2        x    x_catB  catB 0.115273608 0.29380616                 2
+ #  3        z   z_clean clean 0.237601767 0.13176020                 0
+ #  4        z   z_isBAD isBAD 0.296065432 0.09248399                 0
+ #  5        x  x_lev_NA   lev 0.296065432 0.09248399                 0
+ #  6        x x_lev_x.a   lev 0.130005705 0.26490379                 0
+ #  7        x x_lev_x.b   lev 0.006067337 0.80967242                 0
 
 # help("prepare")
 
@@ -159,27 +159,27 @@ dTrainCTreated <- prepare(treatmentsC,dTrainC,pruneSig=1.0,scale=TRUE)
 varsC <- setdiff(colnames(dTrainCTreated),'y')
 # all input variables should be mean 0
 sapply(dTrainCTreated[,varsC,drop=FALSE],mean)
- #       x_lev_NA     x_lev_x.a     x_lev_x.b        x_catP        x_catB 
- #   3.965082e-18 -1.982154e-17  9.917546e-19  1.585994e-16  0.000000e+00 
- #        z_clean       z_isBAD 
- #   7.927952e-18 -7.926292e-18
+ #         x_catP        x_catB       z_clean       z_isBAD      x_lev_NA 
+ #   1.585994e-16  0.000000e+00  7.927952e-18 -7.926292e-18  3.965082e-18 
+ #      x_lev_x.a     x_lev_x.b 
+ #  -1.982154e-17  9.917546e-19
 # all non NA slopes should be 1
 sapply(varsC,function(c) { lm(paste('y',c,sep='~'),
    data=dTrainCTreated)$coefficients[[2]]})
- #   x_lev_NA x_lev_x.a x_lev_x.b    x_catP    x_catB   z_clean   z_isBAD 
+ #     x_catP    x_catB   z_clean   z_isBAD  x_lev_NA x_lev_x.a x_lev_x.b 
  #          1         1         1         1         1         1         1
 dTestCTreated <- prepare(treatmentsC,dTestC,pruneSig=c(),scale=TRUE)
 print(dTestCTreated)
- #      x_lev_NA  x_lev_x.a   x_lev_x.b     x_catP     x_catB  z_clean
[…]

[…]
