### 2018-02-16 71faee2
work on parallelism
