+++ extras/PerfTests.Rmd
@@ -7,8 +7,9 @@ output: github_document
 
 ```{r}
 library("vtreat")
+packageVersion("vtreat")
 
-useParallel <- FALSE
+useParallel <- TRUE
 
 mkEx <- function(n_rows, 
                  n_cat_columns, n_num_columns, n_irrel_columns,
@@ -58,7 +59,7 @@ if(useParallel) {
   ncores <- parallel::detectCores()
   parallelCluster <- parallel::makeCluster(ncores)
 }
-n_rows <- 100000
+n_rows <- 500000
 ```
 
 Get a base timing of a moderately large task.
