+++ extras/PerfTests2.md
@@ -61,7 +61,7 @@ if(useParallel) {
   ncores <- parallel::detectCores()
   parallelCluster <- parallel::makeCluster(ncores)
 }
-n_rows <- 500000
+n_rows <- 2000000
 ```
 
 Get a base timing of a moderately large task.
@@ -87,7 +87,7 @@ system.time(
 ```
 
     ##    user  system elapsed 
-    ##  19.499   2.802  72.815
+    ## 168.960  18.400 349.123
 
 ``` r
 knitr::kable(tplan$treatments$scoreFrame)
@@ -95,22 +95,22 @@ knitr::kable(tplan$treatments$scoreFrame)
 
 | varName                       | varMoves |      rsq|        sig| needsSplit |  extraModelDegrees| origName    | code  |
 |:------------------------------|:---------|--------:|----------:|:-----------|------------------:|:------------|:------|
-| var\_cat\_1\_catP             | TRUE     |  2.1e-06|  0.2271728| TRUE       |                  4| var\_cat\_1 | catP  |
-| var\_cat\_1\_catB             | TRUE     |  1.0e-07|  0.8040180| TRUE       |                  4| var\_cat\_1 | catB  |
-| var\_cat\_2\_catP             | TRUE     |  0.0e+00|  0.9598669| TRUE       |                  4| var\_cat\_2 | catP  |
-| var\_cat\_2\_catB             | TRUE     |  2.0e-06|  0.2338736| TRUE       |                  4| var\_cat\_2 | catB  |
-| var\_num\_1\_clean            | TRUE     |  6.0e-07|  0.5322812| FALSE      |                  0| var\_num\_1 | clean |
-| var\_num\_2\_clean            | TRUE     |  4.4e-06|  0.0808018| FALSE      |                  0| var\_num\_2 | clean |
-| var\_cat\_1\_lev\_x.lev\_a\_1 | TRUE     |  4.0e-07|  0.5872282| FALSE      |                  0| var\_cat\_1 | lev   |
-| var\_cat\_1\_lev\_x.lev\_a\_2 | TRUE     |  2.1e-06|  0.2264627| FALSE      |                  0| var\_cat\_1 | lev   |
-| var\_cat\_1\_lev\_x.lev\_a\_3 | TRUE     |  2.9e-06|  0.1557756| FALSE      |                  0| var\_cat\_1 | lev   |
-| var\_cat\_1\_lev\_x.lev\_a\_4 | TRUE     |  0.0e+00|  0.9102993| FALSE      |                  0| var\_cat\_1 | lev   |
-| var\_cat\_1\_lev\_x.lev\_a\_5 | TRUE     |  1.0e-07|  0.8274104| FALSE      |                  0| var\_cat\_1 | lev   |
-| var\_cat\_2\_lev\_x.lev\_a\_1 | TRUE     |  1.8e-06|  0.2582663| FALSE      |                  0| var\_cat\_2 | lev   |
-| var\_cat\_2\_lev\_x.lev\_a\_2 | TRUE     |  2.0e-07|  0.7436009| FALSE      |                  0| var\_cat\_2 | lev   |
-| var\_cat\_2\_lev\_x.lev\_a\_3 | TRUE     |  8.0e-07|  0.4586753| FALSE      |                  0| var\_cat\_2 | lev   |
-| var\_cat\_2\_lev\_x.lev\_a\_4 | TRUE     |  2.1e-06|  0.2273941| FALSE      |                  0| var\_cat\_2 | lev   |
[…]
