### 2018-02-17 c7143dc
show split method
