+++ extras/LargeExample.md
@@ -1,0 +1,132 @@
+Large Example
+================
+Win-Vector LLC
+2/17/2018
+
+[`vtreat`](https://winvector.github.io/vtreat/) is a an [`R`](https://cran.r-project.org) 'data.frame' processor/conditioner that prepares real-world data for predictive modeling in a statistically sound manner. `vtreat` prepares variables so that data has fewer exceptional cases, making it easier to safely use models in production. Common problems `vtreat` defends against: 'Inf', 'NA', too many categorical levels, rare categorical levels, and new categorical levels (levels seen during application, but not during training). Reference: ["`vtreat`: a data.frame Processor for Predictive Modeling", Zumel, Mount, 2016](https://arxiv.org/abs/1611.09477), [DOI:10.5281/zenodo.1173314](https://doi.org/10.5281/zenodo.1173314).
+
+[`vtreat::mkCrossFrameCExperiment()`](https://winvector.github.io/vtreat/reference/mkCrossFrameCExperiment.html) and [`vtreat::mkCrossFrameCExperiment()`](https://winvector.github.io/vtreat/reference/mkCrossFrameNExperiment.html) are the most statistically efficient methods `vtreat` supplies. So we often advise them, especially for data with few rows. However they have non-trivial computational cost. For data with very many rows we suggest the more computationally efficient [`vtreat`::designTreamentsC()](https://winvector.github.io/vtreat/reference/designTreatmentsC.html) and [`vtreat`::designTreatmensN()](https://winvector.github.io/vtreat/reference/designTreatmentsN.html).
+
+Here is an example (based on [`vtreat issue 12`](https://github.com/WinVector/vtreat/issues/12), [perf tests](https://github.com/WinVector/vtreat/blob/master/extras/PerfTests.md), and [perf tests2](https://github.com/WinVector/vtreat/blob/parallel_ind/extras/PerfTests2.md)).
+
+Example data (in this case no variables are truly related to the outcome to be predicted).
+
+``` r
+library("vtreat")
+packageVersion("vtreat")
+```
+
+    ## [1] '1.0.3'
+
+``` r
+useParallel <- TRUE
+
+mkEx <- function(n_rows, 
+                 n_cat_columns, n_num_columns, n_irrel_columns,
+                 n_cat_levels_a, n_cat_levels_b) {
+  n_cols <- n_cat_columns + n_num_columns + n_irrel_columns + 2
+  d <- as.data.frame(matrix(data = rnorm(n_rows * n_cols), 
+                            nrow = n_rows, ncol = n_cols))
+  cat_names <- NULL
+  num_names <- NULL
+  irrel_names <- NULL
+  if(n_cat_columns>0) {
+    cat_names <- paste0('var_cat_', seq_len(n_cat_columns))
+  }
+  if(n_num_columns>0) {
+    num_names <- paste0('var_num_', seq_len(n_num_columns))
+  }
+  if(n_irrel_columns>0) {
+    irrel_names <- paste0('irrel_', seq_len(n_irrel_columns))
+  }
+  y_names <- c("yC", "yN")
+  colnames(d) <- c(cat_names, num_names, irrel_names, y_names)
+  d$yC <- ifelse(d$yC>=0, "Y", "N")
+  levels_a <- paste0("lev_a_", seq_len(n_cat_levels_a))
+  levels_b <- NULL
+  if(n_cat_levels_b>0) {
+    levels_b <- paste0("lev_b_", seq_len(n_cat_levels_b))
+  }
+  for(ci in cat_names) {
+    a_set <- rep(TRUE, n_rows)
+    if(n_cat_levels_b>0) {
+      a_set <- runif(n_rows)>=0.5
+    }
[…]
