### 2018-02-17 811bec5
add novel value warnings
