+++ NEWS.md
@@ -2,6 +2,7 @@
 # vtreat 1.0.3 2017/02/17
 
   * parallel indicator calculation.
+  * add optional warning on encountering novel levels in prepare.
 
 # vtreat 1.0.2 2017/01/20
 
