+++ NAMESPACE
@@ -12,9 +12,11 @@ export(kWayStratifiedY)
 export(makekWayCrossValidationGroupedByColumn)
 export(mkCrossFrameCExperiment)
 export(mkCrossFrameNExperiment)
+export(novel_value_summary)
 export(oneWayHoldout)
 export(prepare)
 export(problemAppPlan)
+export(track_values)
 export(vnames)
 export(vorig)
 importFrom(stats,aggregate)

[…]
