+++ man/mkCrossFrameCExperiment.Rd
@@ -9,7 +9,7 @@ mkCrossFrameCExperiment(dframe, varlist, outcomename, outcometarget, ...,
   rareSig = 1, collarProb = 0, codeRestriction = NULL,
   customCoders = NULL, scale = FALSE, doCollar = FALSE,
   splitFunction = NULL, ncross = 3, forceSplit = FALSE,
-  catScaling = FALSE, parallelCluster = NULL)
+  catScaling = FALSE, verbose = TRUE, parallelCluster = NULL)
 }
 \arguments{
 \item{dframe}{Data frame to learn treatments from (training data), must have at least 1 row.}
@@ -49,6 +49,8 @@ mkCrossFrameCExperiment(dframe, varlist, outcomename, outcometarget, ...,
 \item{forceSplit}{logical, if TRUE force cross-validated significance calculatons on all variables.}
 
 \item{catScaling}{optional, if TRUE use glm() linkspace, if FALSE use lm() for scaling.}
+
+\item{verbose}{if TRUE print progress.}
 
 \item{parallelCluster}{(optional) a cluster object created by package parallel or package snow}
 }
