+++ README.md
@@ -145,8 +145,8 @@ treatmentsC <- designTreatmentsC(dTrainC,colnames(dTrainC),'y',TRUE,
                                  verbose=FALSE)
 print(treatmentsC$scoreFrame[,c('origName', 'varName', 'code', 'rsq', 'sig', 'extraModelDegrees')])
  #    origName   varName  code         rsq        sig extraModelDegrees
- #  1        x    x_catP  catP 0.059315943 0.45141252                 2
- #  2        x    x_catB  catB 0.029824903 0.59334713                 2
+ #  1        x    x_catP  catP 0.057741424 0.45748159                 2
+ #  2        x    x_catB  catB 0.019483838 0.66603146                 2
  #  3        z   z_clean clean 0.237601767 0.13176020                 0
  #  4        z   z_isBAD isBAD 0.296065432 0.09248399                 0
  #  5        x  x_lev_NA   lev 0.296065432 0.09248399                 0
@@ -192,9 +192,9 @@ treatmentsN = designTreatmentsN(dTrainN,colnames(dTrainN),'y',
                                 verbose=FALSE)
 print(treatmentsN$scoreFrame[,c('origName', 'varName', 'code', 'rsq', 'sig', 'extraModelDegrees')])
  #    origName   varName  code          rsq       sig extraModelDegrees
- #  1        x    x_catP  catP 3.558824e-01 0.1184999                 2
- #  2        x    x_catN  catN 2.663337e-02 0.6993947                 2
- #  3        x    x_catD  catD 4.226472e-02 0.6252538                 2
+ #  1        x    x_catP  catP 2.197309e-01 0.2413478                 2
+ #  2        x    x_catN  catN 7.286735e-02 0.5179131                 2
+ #  3        x    x_catD  catD 2.227248e-01 0.2377286                 2
  #  4        z   z_clean clean 2.880952e-01 0.1701892                 0
  #  5        z   z_isBAD isBAD 3.333333e-01 0.1339746                 0
  #  6        x  x_lev_NA   lev 3.333333e-01 0.1339746                 0
