### 2018-02-19 5583690
add replacement sampler
