+++ README.md
@@ -144,14 +144,14 @@ dTestC <- data.frame(x=c('a','b','c',NA),z=c(10,20,30,NA))
 treatmentsC <- designTreatmentsC(dTrainC,colnames(dTrainC),'y',TRUE,
                                  verbose=FALSE)
 print(treatmentsC$scoreFrame[,c('origName', 'varName', 'code', 'rsq', 'sig', 'extraModelDegrees')])
- #    origName   varName  code         rsq        sig extraModelDegrees
- #  1        x    x_catP  catP 0.057741424 0.45748159                 2
- #  2        x    x_catB  catB 0.019483838 0.66603146                 2
- #  3        z   z_clean clean 0.237601767 0.13176020                 0
- #  4        z   z_isBAD isBAD 0.296065432 0.09248399                 0
- #  5        x  x_lev_NA   lev 0.296065432 0.09248399                 0
- #  6        x x_lev_x.a   lev 0.130005705 0.26490379                 0
- #  7        x x_lev_x.b   lev 0.006067337 0.80967242                 0
+ #    origName   varName  code          rsq        sig extraModelDegrees
+ #  1        x    x_catP  catP 1.559780e-01 0.22202097                 2
+ #  2        x    x_catB  catB 1.142159e-05 0.99166241                 2
+ #  3        z   z_clean clean 2.376018e-01 0.13176020                 0
+ #  4        z   z_isBAD isBAD 2.960654e-01 0.09248399                 0
+ #  5        x  x_lev_NA   lev 2.960654e-01 0.09248399                 0
+ #  6        x x_lev_x.a   lev 1.300057e-01 0.26490379                 0
+ #  7        x x_lev_x.b   lev 6.067337e-03 0.80967242                 0
 
 # help("prepare")
 
@@ -192,9 +192,9 @@ treatmentsN = designTreatmentsN(dTrainN,colnames(dTrainN),'y',
                                 verbose=FALSE)
 print(treatmentsN$scoreFrame[,c('origName', 'varName', 'code', 'rsq', 'sig', 'extraModelDegrees')])
  #    origName   varName  code          rsq       sig extraModelDegrees
- #  1        x    x_catP  catP 2.197309e-01 0.2413478                 2
- #  2        x    x_catN  catN 7.286735e-02 0.5179131                 2
- #  3        x    x_catD  catD 2.227248e-01 0.2377286                 2
+ #  1        x    x_catP  catP 1.764706e-01 0.3001022                 2
+ #  2        x    x_catN  catN 3.533569e-03 0.8888051                 2
+ #  3        x    x_catD  catD 3.743113e-01 0.1069707                 2
  #  4        z   z_clean clean 2.880952e-01 0.1701892                 0
  #  5        z   z_isBAD isBAD 3.333333e-01 0.1339746                 0
  #  6        x  x_lev_NA   lev 3.333333e-01 0.1339746                 0
