+++ NAMESPACE
@@ -9,6 +9,7 @@ export(designTreatmentsZ)
 export(getSplitPlanAppLabels)
 export(kWayCrossValidation)
 export(kWayStratifiedY)
+export(kWayStratifiedYReplace)
 export(makekWayCrossValidationGroupedByColumn)
 export(mkCrossFrameCExperiment)
 export(mkCrossFrameNExperiment)
