### 2018-02-24 #584 @earino
@JohnMount greeting! I was going through the review checklist, and most things are very straightforward. I should add that this is my first review for JOSS, so I please bare with me. I however couldn't verify:

- community guidelines
- all archival references that should have a DOI list one (e.g., papers, datasets, software)?

If you could help me find them, I would appreciate it!
