### 2018-02-24 #584 @JohnMount
@earino Thanks for volunteering!  Please don't hesitate to ask me questions.

 * The project has not yet added explicit community guidelines.  We certainly must adhere to the https://help.github.com/articles/github-community-guidelines/ by being on GitHub and subject to their terms of service.  I'll can look into finding some explicit guidelines if that is a requirement (or is it enough to note the project does or does not have some?).
 * DOIs, well part of this submission is to help get a DOI for this work.  But we do have some relevant DOIs: https://zenodo.org/record/1173314#.WpHTaWaZNqc https://zenodo.org/record/1173318#.WpHTZ2aZNqc .
