### 2018-02-24 #584 @earino
@JohnMount I'm happy to. I've spoken about this package in a few talks, so it seems like a good intersection of interest and aptitutde :)
 - I looked a number of other issues, and they all seem to have a CONTRIBUTING.md. Some of them are longer than others, [here](https://github.com/piermorel/gramm/blob/master/CONTRIBUTING.md)'s an example. I think that the expectation is that the expectation is explicit?
- I think you're supposed to include those DOIs in the references section of paper.md? Maybe I'm wrong on this one.
