### 2018-02-24 #584 @JohnMount
I'll run down a CONTRIBUTING.md right now.  I just tried adding DOIs to paper.bib, but the re-render seemed to ignore them. If you know how to get them to render that would help a bit.  I am going to also add the DOIs to the README.md so they will be there in a little bit.
