### 2018-02-24 #584 @JohnMount
Ah my mistake.  I needed to merge the branch in to get the DOIs in.  I now have DOIs in the paper.bib and the re-generated paper shows them. I'll add the paper one to the README.md (it doesn't make sense for the repository to refer to the code snapshot).  And I'll add a CONTRIBUTING.md .
