### 2018-02-24 be2caa2
add contributing and update
