+++ CONTRIBUTING.md
@@ -1,0 +1,8 @@
+# Contributing
+
+The `vtreat` project is happy to examine good faith issues and pull requests [here](https://github.com/WinVector/vtreat).
+Also please feel free to email the maintaining contributor: John Mount `jmount@win-vector.com`.
+
+Please understand we can not accept any code without having full rights and license assigned to Win-Vector LLC.
+
+Also we expect contributors to follow [GitHub Community Guidelines](https://help.github.com/articles/github-community-guidelines/).
