### 2018-02-24 #584 @JohnMount
Okay, I think I have it all fixed up now. DOIs in the [paper](https://github.com/openjournals/joss-papers/blob/joss.00584/joss.00584/10.21105.joss.00584.pdf) (now visible), and in [README.md](https://github.com/WinVector/vtreat/blob/master/README.md).  Now have a simple [CONTRIBUTING.md](https://github.com/WinVector/vtreat/blob/master/CONTRIBUTING.md) file.
