### 2018-03-10 e2fe8e2
prepare for CRAN release
