+++ README.md
@@ -144,14 +144,14 @@ dTestC <- data.frame(x=c('a','b','c',NA),z=c(10,20,30,NA))
 treatmentsC <- designTreatmentsC(dTrainC,colnames(dTrainC),'y',TRUE,
                                  verbose=FALSE)
 print(treatmentsC$scoreFrame[,c('origName', 'varName', 'code', 'rsq', 'sig', 'extraModelDegrees')])
- #    origName   varName  code          rsq        sig extraModelDegrees
- #  1        x    x_catP  catP 1.030137e-01 0.32099590                 2
- #  2        x    x_catB  catB 1.125399e-05 0.99172381                 2
- #  3        z   z_clean clean 2.376018e-01 0.13176020                 0
- #  4        z   z_isBAD isBAD 2.960654e-01 0.09248399                 0
- #  5        x  x_lev_NA   lev 2.960654e-01 0.09248399                 0
- #  6        x x_lev_x.a   lev 1.300057e-01 0.26490379                 0
- #  7        x x_lev_x.b   lev 6.067337e-03 0.80967242                 0
+ #    origName   varName  code         rsq        sig extraModelDegrees
+ #  1        x    x_catP  catP 0.285095342 0.09874390                 2
+ #  2        x    x_catB  catB 0.117443640 0.28930668                 2
+ #  3        z   z_clean clean 0.237601767 0.13176020                 0
+ #  4        z   z_isBAD isBAD 0.296065432 0.09248399                 0
+ #  5        x  x_lev_NA   lev 0.296065432 0.09248399                 0
+ #  6        x x_lev_x.a   lev 0.130005705 0.26490379                 0
+ #  7        x x_lev_x.b   lev 0.006067337 0.80967242                 0
 
 # help("prepare")
 
@@ -192,9 +192,9 @@ treatmentsN = designTreatmentsN(dTrainN,colnames(dTrainN),'y',
                                 verbose=FALSE)
 print(treatmentsN$scoreFrame[,c('origName', 'varName', 'code', 'rsq', 'sig', 'extraModelDegrees')])
  #    origName   varName  code          rsq       sig extraModelDegrees
- #  1        x    x_catP  catP 1.894619e-01 0.2810898                 2
- #  2        x    x_catN  catN 2.532154e-01 0.2036653                 2
- #  3        x    x_catD  catD 4.512437e-02 0.6135229                 2
+ #  1        x    x_catP  catP 1.304740e-01 0.3793327                 2
+ #  2        x    x_catN  catN 2.616674e-01 0.1950536                 2
+ #  3        x    x_catD  catD 9.705853e-02 0.4525551                 2
  #  4        z   z_clean clean 2.880952e-01 0.1701892                 0
  #  5        z   z_isBAD isBAD 3.333333e-01 0.1339746                 0
  #  6        x  x_lev_NA   lev 3.333333e-01 0.1339746                 0
