+++ cran-comments.md
@@ -1,22 +1,28 @@
 
+Maintinance release with parallel performance fixes.
+
 ## Test environments
 
  * Windows
- * using R Under development (unstable) (2018-01-19 r74138)
- * using platform: x86_64-w64-mingw32 (64-bit)
+
  
  * OSX
  * using R version 3.4.3 (2017-11-30)
  * using platform: x86_64-apple-darwin15.6.0 (64-bit)
 
 
-## R CMD check --as-cran vtreat_1.0.2.tar.gz
+## R CMD check --as-cran vtreat_1.0.3.tar.gz
 
-* this is package ‘vtreat’ version ‘1.0.2’
-* checking CRAN incoming feasibility ... Note_to_CRAN_maintainers
-Maintainer: ‘John Mount <jmount@win-vector.com>’
+  * using option ‘--as-cran’
+  * checking for file ‘vtreat/DESCRIPTION’ ... OK
+  * checking extension type ... Package
+  * this is package ‘vtreat’ version ‘1.0.3’
+  * checking CRAN incoming feasibility ... Note_to_CRAN_maintainers
+    Maintainer: ‘John Mount <jmount@win-vector.com>’
 
 Status: OK
+
+No other notes, warnings or errors.
 
 
 ## Downstream dependencies
