### 2018-03-10 da5fa2e
prepare for CRAN
