+++ cran-comments.md
@@ -4,7 +4,14 @@ Maintinance release with parallel performance fixes.
 ## Test environments
 
  * Windows
-
+ * using R Under development (unstable) (2018-03-09 r74376)
+ * using platform: x86_64-w64-mingw32 (64-bit)
+ 
+  Status: 1 NOTE
+  Possibly mis-spelled words in DESCRIPTION:
+    Zumel (20:63)
+    vtreat (20:6)
+  (not mis-spelled and now marked with single quotes)
  
  * OSX
  * using R version 3.4.3 (2017-11-30)
@@ -31,3 +38,6 @@ No declared reverse dependencies:
 
      devtools::revdep('vtreat')
      character(0)
+     
+Zumel is not mis-spelling.
+
