+++ DESCRIPTION
@@ -16,7 +16,7 @@ Description: A 'data.frame' processor/conditioner that prepares real-world data 
     it easier to safely use models in production. Common problems 'vtreat' defends
     against: 'Inf', 'NA', too many categorical levels, rare categorical levels, and new
     categorical levels (levels seen during application, but not during training). Reference: 
-    "vtreat: a data.frame Processor for Predictive Modeling", Zumel, Mount, 2016, DOI:10.5281/zenodo.1173314.
+    "'vtreat': a data.frame Processor for Predictive Modeling", 'Zumel', 'Mount', 2016, DOI:10.5281/zenodo.1173314.
 License: GPL-3
 Imports:
     stats
