### 2018-03-12 #584 @JohnMount
Thanks for accepting the paper.  The article seems to be a dead link even though the check looks good on the thread.  http://joss.theoj.org/papers/10.21105/joss.00584
