### 2018-03-12 #584 @JohnMount
The paper was there for a little while and then disappeared again.

 https://github.com/openjournals/joss-reviews/issues/584 (paper does not show)

https://www.theoj.org/joss-papers/joss.00584/10.21105.joss.00584.pdf (broken link)
