### 2018-03-12 #584 @JohnMount
Looks like it disappeared again. Does JOSS get the paper from the tar file?  Because it is not there- I used the CRAN distribution to build the marked release of the software (not the raw Github account).

If that is the problem we can up the DOI to have the full GitHub contents:

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1196652.svg)](https://doi.org/10.5281/zenodo.1196652)

John
