### 2018-03-12 #584 @arfon
Looks like it's back again. I think this is due to the GitHub pages outage earlier.

> Does JOSS get the paper from the tar file?

No, we pull this straight from GitHub: https://github.com/openjournals/joss-papers/blob/a516dc7ddb45a1cbc0eabe7ecd9446dc2d657256/joss.00584/10.21105.joss.00584.pdf
